+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Changes
 - ([\#71](https://github.com/forbole/juno/pull/71)) Retry RPC client connection upon failure instead of panic
 - ([\#72](https://github.com/forbole/juno/pull/72)) Updated missing blocks parsing 
+- ([\#73](https://github.com/forbole/juno/pull/73)) Re-enqueue failed block after average block time
 
 ## v3.3.0
 ### Changes